Refactor indexer - implement more robust methods
