+++ package-manager-indexer/README.md
@@ -13,7 +13,7 @@ This service listens to events from the AutonomysPackageRegistry smart contract 
 
 ## Prerequisites
 
-- Node.js v18+
+- Node.js v20+
 - PostgreSQL 14+
 - Access to an Ethereum-compatible RPC endpoint
 
@@ -147,15 +147,3 @@ The indexer service follows a modular design:
 3. **Event Handlers**: Process blockchain events and update the database
 4. **API Server**: Exposes HTTP endpoints for querying the database
 5. **Migration System**: Manages database schema changes
-
-## Contributing
-
-1. Fork the repository
-2. Create your feature branch: `git checkout -b feature/my-new-feature`
-3. Commit your changes: `git commit -am 'Add some feature'`
-4. Push to the branch: `git push origin feature/my-new-feature`
-5. Submit a pull request
-
-## License
-
-This project is licensed under the MIT License - see the LICENSE file for details.
