+++ package-manager-indexer/package.json
@@ -15,6 +15,7 @@
     "reset-database": "node dist/scripts/resetDatabase.js"
   },
   "dependencies": {
+    "@autonomys/auto-dag-data": "^1.4.18",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
