+++ package-manager-indexer/src/index.ts
@@ -2,152 +2,21 @@ import { createLogger } from './utils/logger.js';
 import { config } from './config/index.js';
 import { 
   initContract, 
-  setupEventListeners, 
-  processHistoricalEvents,
-  ToolRegisteredEvent, 
-  ToolUpdatedEvent, 
-  OwnershipTransferredEvent,
-  parseVersion
+  startIndexing,
+  EventCallbacks
 } from './blockchain/contractService.js';
 import {
-  saveTool,
-  getToolByNameHash,
-  saveToolVersion,
-  updateToolOwner,
-  updateToolVersionMetadata,
   getLastProcessedBlock,
-  updateLastProcessedBlock
 } from './db/repositories/toolRepository.js';
-import { ethers } from 'ethers';
 import { startServer } from './api/server.js';
+import {
+  handleToolRegistered,
+  handleToolUpdated,
+  handleOwnershipTransferred,
+  handleProcessedBlock
+} from './handlers/eventHandlers.js';
 
 const logger = createLogger('indexer-service');
-
-// Handler for ToolRegistered events
-const handleToolRegistered = async (event: ToolRegisteredEvent): Promise<void> => {
-  logger.info(`Processing ToolRegistered: ${event.name} v${event.major}.${event.minor}.${event.patch}`);
-  
-  try {
-    // Get the name hash
-    const nameHash = ethers.keccak256(ethers.toUtf8Bytes(event.name));
-    
-    // Check if tool already exists
-    const existingTool = await getToolByNameHash(nameHash);
-    
-    let toolId: number;
-    
-    if (existingTool) {
-      logger.info(`Tool ${event.name} already exists, using existing record`);
-      toolId = existingTool.id;
-    } else {
-      // Save the new tool
-      const tool = await saveTool(event.name, nameHash, event.publisher);
-      logger.info(`New tool ${event.name} saved with ID ${tool.id}`);
-      toolId = tool.id;
-    }
-    
-    // Save the tool version
-    const version = parseVersion(event.major, event.minor, event.patch);
-    await saveToolVersion(
-      toolId,
-      version,
-      event.cidHash,
-      event.cidHash, // Initially metadata hash is the same as cid hash
-      event.publisher,
-      event.timestamp
-    );
-    
-    logger.info(`Tool version ${event.major}.${event.minor}.${event.patch} saved for ${event.name}`);
-  } catch (error) {
-    logger.error('Error handling ToolRegistered event:', error);
-    throw error;
-  }
-};
-
-// Handler for ToolUpdated events
-const handleToolUpdated = async (event: ToolUpdatedEvent): Promise<void> => {
-  logger.info(`Processing ToolUpdated: ${event.name} v${event.major}.${event.minor}.${event.patch}`);
-  
-  try {
-    // Get the name hash
-    const nameHash = ethers.keccak256(ethers.toUtf8Bytes(event.name));
-    
-    // Get the tool
-    const tool = await getToolByNameHash(nameHash);
-    
-    if (!tool) {
-      // This shouldn't happen - we should have received a ToolRegistered event first
-      logger.warn(`Tool ${event.name} not found, creating it now`);
-      const newTool = await saveTool(event.name, nameHash, event.publisher);
-      
-      // Save the version
-      const version = parseVersion(event.major, event.minor, event.patch);
-      await saveToolVersion(
-        newTool.id,
-        version,
-        event.cidHash,
-        event.cidHash, // Initially metadata hash is the same as cid hash  
-        event.publisher,
-        event.timestamp
-      );
-    } else {
-      // Check if this is a metadata update or a new version
-      const version = parseVersion(event.major, event.minor, event.patch);
-      
-      // Try to update the metadata first
-      const updated = await updateToolVersionMetadata(tool.id, version, event.cidHash);
-      
-      if (!updated) {
-        // If update failed, it means this is a new version
-        await saveToolVersion(
-          tool.id,
-          version,
-          event.cidHash,
-          event.cidHash, // Initially metadata hash is the same as cid hash
-          event.publisher,
-          event.timestamp
-        );
-        logger.info(`New tool version ${event.major}.${event.minor}.${event.patch} saved for ${event.name}`);
-      } else {
-        logger.info(`Updated metadata for ${event.name} v${event.major}.${event.minor}.${event.patch}`);
-      }
-    }
-  } catch (error) {
-    logger.error('Error handling ToolUpdated event:', error);
-    throw error;
-  }
-};
-
-// Handler for OwnershipTransferred events
-const handleOwnershipTransferred = async (event: OwnershipTransferredEvent): Promise<void> => {
-  logger.info(`Processing OwnershipTransferred: ${event.name} from ${event.previousOwner} to ${event.newOwner}`);
-  
-  try {
-    // Get the name hash
-    const nameHash = ethers.keccak256(ethers.toUtf8Bytes(event.name));
-    
-    // Update the tool owner
-    const updated = await updateToolOwner(nameHash, event.newOwner);
-    
-    if (updated) {
-      logger.info(`Ownership of ${event.name} transferred to ${event.newOwner}`);
-    } else {
-      logger.warn(`Tool ${event.name} not found for ownership transfer`);
-    }
-  } catch (error) {
-    logger.error('Error handling OwnershipTransferred event:', error);
-    throw error;
-  }
-};
-
-// Handler for block processed events
-const handleProcessedBlock = async (blockNumber: number): Promise<void> => {
-  try {
-    await updateLastProcessedBlock(blockNumber);
-  } catch (error) {
-    logger.error('Error updating last processed block:', error);
-  }
-};
 
 // Start the indexer
 const startIndexer = async (): Promise<void> => {
@@ -164,7 +33,7 @@ const startIndexer = async (): Promise<void> => {
     const { contract } = initContract();
 
     // Create the event callbacks
-    const eventCallbacks = {
+    const eventCallbacks: EventCallbacks = {
       onToolRegistered: handleToolRegistered,
       onToolUpdated: handleToolUpdated,
       onOwnershipTransferred: handleOwnershipTransferred,
@@ -175,21 +44,9 @@ const startIndexer = async (): Promise<void> => {
     const apiPort = process.env.API_PORT ? parseInt(process.env.API_PORT) : 3000;
     const { shutdown: shutdownApi } = startServer(apiPort);
     
-    // First process historical events
-    logger.info('Processing historical events...');
-    const lastProcessedHistoricalBlock = await processHistoricalEvents(
-      contract,
-      startBlock,
-      eventCallbacks
-    );
-    
-    // Then setup event listeners for new events
-    logger.info(`Historical processing complete. Now listening for new events from block ${lastProcessedHistoricalBlock + 1}`);
-    const cleanupListeners = setupEventListeners(
-      contract,
-      lastProcessedHistoricalBlock + 1,
-      eventCallbacks
-    );
+    // Start processing events using the unified startIndexing function
+    logger.info('Starting event processing...');
+    const cleanupListeners = await startIndexing(contract, startBlock, eventCallbacks);
     
     logger.info('Indexer started successfully');
